+++ meta-scene/importers/pbrt/import_pbrt_integrator.cpp
@@ -15,7 +15,7 @@ namespace meta_scene::importers::pbrt {
 				auto [type, name] = context.peek_type_and_name();
 
 				if (type == PBRT_INTEGER_TOKEN) {
-					const auto value = context.peek_integer<size_t>();
+					const auto value = context.peek_integer<uint32>();
 					
 					if (name == "maxdepth") META_SCENE_FINISHED_AND_RETURN(integrator.depth = value);
 				}
@@ -41,7 +41,7 @@ namespace meta_scene::importers::pbrt {
 				auto [type, name] = context.peek_type_and_name();
 
 				if (type == PBRT_INTEGER_TOKEN) {
-					const auto value = context.peek_integer<size_t>();
+					const auto value = context.peek_integer<uint32>();
 
 					if (name == "maxdepth") META_SCENE_FINISHED_AND_RETURN(integrator.depth = value);
 				}
